# doc/swagger/API ExchangeRate.swagger.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["auth"]}},"/exchange-rate/calculate":{"post":{"operationId":"ExchangeRateController_calculate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalculateRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalculateResponseDto"}}}}},"tags":["exchange-rate"],"security":[{"bearer":[]}]}},"/exchange-rate/update":{"post":{"operationId":"ExchangeRateController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExchangeRate"}}}}},"tags":["exchange-rate"],"security":[{"bearer":[]}]}}},"info":{"title":"Exchange Rate API","description":"Reto técnico de Wally","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000","description":"Local environment"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"CalculateRequestDto":{"type":"object","properties":{"amount":{"type":"number"},"originCurrency":{"type":"string"},"destinationCurrency":{"type":"string"}},"required":["amount","originCurrency","destinationCurrency"]},"CalculateResponseDto":{"type":"object","properties":{"amount":{"type":"number"},"originCurrency":{"type":"string"},"destinationCurrency":{"type":"string"},"rate":{"type":"number"},"convertedAmount":{"type":"number"}},"required":["amount","originCurrency","destinationCurrency","rate","convertedAmount"]},"UpdateRequestDto":{"type":"object","properties":{"originCurrency":{"type":"string"},"destinationCurrency":{"type":"string"},"rate":{"type":"number"}},"required":["originCurrency","destinationCurrency","rate"]},"ExchangeRate":{"type":"object","properties":{}}}}}